+++ python/image/training/image_classification/output/log/training_log.csv
@@ -1,0 +1,4 @@
+epoch,accuracy,loss,val_accuracy,val_loss
+0,0.05451807379722595,6.40269660949707,0.0060240961611270905,19.633617401123047
+1,0.09578312933444977,5.3788981437683105,0.02530120499432087,17.949983596801758
+2,0.33283132314682007,2.82672381401062,0.0060240961611270905,38.66039276123047
