--- python/image/training/output/log/training_log.csv
@@ -1,6 +1,0 @@
-epoch,accuracy,loss,val_accuracy,val_loss
-0,0.8990557193756104,0.4745357036590576,0.048406101763248444,28.57962989807129
-1,0.981488823890686,0.07105758786201477,0.5913854837417603,2.6596503257751465
-2,0.9868519306182861,0.04933649301528931,0.7600781321525574,1.2535709142684937
-3,0.9900912046432495,0.03624411299824715,0.7554576992988586,1.3878000974655151
-4,0.9922685623168945,0.030713791027665138,0.6913917064666748,1.7450745105743408
